edit resources and models

+++ resources/customer.py
@@ -1,68 +1,35 @@
-from flask import jsonify, request
+from flask import jsonify, request, make_response
 from flask_restful import Resource
 from flask_jwt_extended import jwt_required, create_access_token
 from mongoengine.errors import DoesNotExist, ValidationError
 from models.customer import Customer
 from models.user import User
 
-class SingleCustomer(Resource):
-    
-    def get(self, customer_id=None):
-        if customer_id:
-            try:
-                customer = Customer.objects.get(id=customer_id)
-            except (DoesNotExist, ValidationError):
-                return {"message": "Customer not found"}, 404
-            return jsonify(customer)
-        else:
-            customers = Customer.objects.all()
-            return jsonify(customers)
-
-    @jwt_required()
-    def post(self):
-        data = request.get_json()
-        name = data['name']
-        email = data['email']
-        password = data['password']
-        is_customer = data['is_customer']
-        is_nail_tech = data['is_nail_tech']
-
-        existing_user = User.objects(email=email).first()
-        if existing_user:
-            return {'message': 'User with email already exists.'}, 409
-
-        if is_customer:
-            customer = Customer(name=name, email=email, password=password, is_customer=is_customer)
-            customer.save()
-            access_token = create_access_token(identity=str(customer.id))
-        else:
-            user = User(name=name, email=email, password=password, is_customer=is_customer, is_nail_tech=is_nail_tech)
-            user.save()
-            access_token = create_access_token(identity=str(user.id))
-
-        # Return the access token along with the success message
-        return {'message': 'User created successfully.', 'access_token': access_token}, 201
-
-    @jwt_required()
-    def put(self, customer_id):
-        try:
-            customer = Customer.objects.get(id=customer_id)
-        except (DoesNotExist, ValidationError):
-            return {"message": "Customer not found"}, 404
-
-        body = request.get_json()
-        customer.update(**body)
-        customer.reload()
+class AllCustomers(Resource):
+    def get(self):
+        customer = Customer.objects.all()
         return jsonify(customer)
 
-    @jwt_required()
-    def delete(self, customer_id):
-        try:
-            customer = Customer.objects.get(id=customer_id)
-        except (DoesNotExist, ValidationError):
-            return {"message": "Customer not found"}, 404
+class SingleCustomer(Resource):
+    def get(self, id):
+        customer = Customer.objects(id=id).first()
+        if customer:
+            return make_response(jsonify(customer=customer))
+        return {"message": "Customer not found"}, 404
 
-        customer.delete()
-        return {"message": "Customer deleted"}
+    def put(self, id):
+        customer = Customer.objects(id=id).first()
+        if customer:
+            body = request.get_json()
+            customer.update(**body)
+            return make_response(jsonify(customer=customer))
+        return {"message": "Customer not found"}, 404
+
+    def delete(self, id):
+        customer = Customer.objects(id=id).first()
+        if customer:
+            customer.delete()
+            return {"message": "Customer deleted"}, 204
+        return {"message": "Customer not found"}, 404
