+++ resources/nailtech.py
@@ -1,4 +1,4 @@
-from flask import jsonify, request
+from flask import jsonify, request, make_response
 from flask_restful import Resource
 from flask_jwt_extended import jwt_required
 from mongoengine.errors import DoesNotExist, ValidationError
@@ -10,43 +10,23 @@ class AllNailTechs(Resource):
         return jsonify(nail_techs)
 
 class SingleNailTech(Resource):
-    
-    def get(self, nail_tech_id=None):
-        if nail_tech_id:
-            try:
-                nail_tech = NailTech.objects.get(id=nail_tech_id)
-            except (DoesNotExist, ValidationError):
-                return {"message": "NailTech not found"}, 404
-            return jsonify(nail_tech)
-        else:
-            nail_techs = NailTech.objects.all()
-            return jsonify(nail_techs)
+    def get(self, id):
+        nail_tech = NailTech.objects(id=id).first()
+        if nail_tech:
+            return make_response(jsonify(nail_tech=nail_tech))
+        return {"message": "NailTech not found"}, 404
 
-    @jwt_required()
-    def post(self):
-        body = request.get_json()
-        nail_tech = NailTech(**body)
-        nail_tech.save()
-        return jsonify(nail_tech)
+    def put(self, id):
+        nail_tech = NailTech.objects(id=id).first()
+        if nail_tech:
+            body = request.get_json()
+            nail_tech.update(**body)
+            return make_response(jsonify(nail_tech=nail_tech))
+        return {"message": "NailTech not found"}, 404
 
-    @jwt_required()
-    def put(self, nail_tech_id):
-        try:
-            nail_tech = NailTech.objects.get(id=nail_tech_id)
-        except (DoesNotExist, ValidationError):
-            return {"message": "NailTech not found"}, 404
-
-        body = request.get_json()
-        nail_tech.update(**body)
-        nail_tech.reload()
-        return jsonify(nail_tech)
-
-    @jwt_required()
-    def delete(self, nail_tech_id):
-        try:
-            nail_tech = NailTech.objects.get(id=nail_tech_id)
-        except (DoesNotExist, ValidationError):
-            return {"message": "NailTech not found"}, 404
-
-        nail_tech.delete()
-        return {"message": "NailTech deletfrom flask import jsonify, request"}
+    def delete(self, id):
+        nail_tech = NailTech.objects(id=id).first()
+        if nail_tech:
+            nail_tech.delete()
+            return {"message": "NailTech deleted"}, 204
+        return {"message": "NailTech not found"}, 404
